pyproject.toml
[tool.poetry]
name = "hotel-booking"
version = "0.1.0"
description = ""
authors = ["Daniil <danokp@mail.ru>"]
readme = "README.md"
packages = [{include = "hotel_booking"}]

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
aiofiles = "^23.2.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.4.0"
flower = "^2.0.1"
pillow = "^10.3.0"
asgiref = "^3.8.1"
sqladmin = {extras = ["full"], version = "^0.17.0"}
python-json-logger = "^2.0.7"
sentry-sdk = {extras = ["fastapi"], version = "^2.6.0"}
fastapi-versioning = "^0.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
ruff = "^0.4.7"
httpx = "^0.27.0"
pytest-asyncio = "0.21.2"
pytest-mock = "^3.14.0"
pyright = "^1.1.368"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "app"
asyncio_mode = "auto"
addopts = "--disable-warnings"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff]
line-length = 89
exclude = [".venv", "app/migrations/*", "app/static/*"]

[tool.ruff.lint.per-file-ignores]
"app/migrations/*" = ["E", "W", "F", "I", "B", "C4", "UP"]
"app/static/*" = ["E", "W", "F", "I", "B", "C4", "UP"]
".venv" = ["E", "W", "F", "I", "B", "C4", "UP"]
"app/*" = ["B008"]
"app/users/dependencies.py" = ["B904"]

[tool.pyright]
include = ["app"]
